+++ development/include/parameters.h
@@ -73,7 +73,7 @@ extern "C" {
          * Send date and time. This is used in Data notification messages.
          */
 #ifdef DLMS_USE_EPOCH_TIME
-        unsigned long time;
+        uint32_t time;
 #else
         struct tm* time;
 #endif //DLMS_USE_EPOCH_TIME
@@ -94,7 +94,7 @@ extern "C" {
         /**
          * Block index.
          */
-        unsigned short blockIndex;
+        uint16_t blockIndex;
         //Serialize data to this PDU.
 #ifdef DLMS_IGNORE_MALLOC
         gxByteBuffer* serializedPdu;
@@ -146,7 +146,7 @@ extern "C" {
          * Send date and time. This is used in Data notification messages.
          */
 #ifdef DLMS_USE_EPOCH_TIME
-        unsigned long time;
+        uint32_t time;
 #else
         struct tm* time;
 #endif // DLMS_USE_EPOCH_TIME
@@ -165,7 +165,7 @@ extern "C" {
         /**
          * Block index.
          */
-        unsigned long blockIndex;
+        uint32_t blockIndex;
         /**
         * Received invoke ID.
         */
