+++ development/src/notify.c
@@ -87,7 +87,7 @@ int notify_addData(
 int notify_generateDataNotificationMessages2(
     dlmsSettings* settings,
 #ifdef DLMS_USE_EPOCH_TIME
-    unsigned long time,
+    uint32_t time,
 #else
     struct tm* time,
 #endif //DLMS_USE_EPOCH_TIME
@@ -118,7 +118,7 @@ int notify_generateDataNotificationMessages2(
 int notify_generateDataNotificationMessages(
     dlmsSettings* settings,
 #ifdef DLMS_USE_EPOCH_TIME
-    unsigned long date,
+    uint32_t date,
 #else
     struct tm* date,
 #endif //DLMS_USE_EPOCH_TIME
@@ -163,7 +163,7 @@ int notify_generateDataNotificationMessages(
 int notify_generatePushSetupMessages(
     dlmsSettings* settings,
 #ifdef DLMS_USE_EPOCH_TIME
-    unsigned long date,
+    uint32_t date,
 #else
     struct tm* date,
 #endif //DLMS_USE_EPOCH_TIME
@@ -273,7 +273,7 @@ int notify_parsePush(
                 return ret;
             }
             index = (unsigned char)var_toInteger(tmp);
-            arr_push(items, key_init(obj, (void*)(unsigned long long)index));
+            arr_push(items, key_init(obj, (void*)(uint64_t)index));
 #endif //DLMS_IGNORE_MALLOC
         }
     }
